Fix initial PATH search, check for cmd built-ins
Change initial search on PATH to first check if a drive or path
component is included in the interpreter name. Such names are only
searched as given (in the current directory). For simple command names,
the PATH is also searched if needed.

If the search fails the cmd built-in commands listed by the `help`
command are searched for a match. Otherwise a "Bad interpreter" error
message is given.

Always change (forward) slashes for backslashes when parsing interpreter
name from shebang line.

Add tests for a path without drive and for a cmd built-in command.